extend hessian_adaptation examples with implied-metric refinement family
